+++ otter/check/notebook.py
@@ -18,8 +18,8 @@ from textwrap import indent
 from urllib.parse import urljoin
 
 from .logs import LogEntry, EventType, Log
-from .utils import grade_zip_file, list_available_tests, resolve_test_info, running_on_colab, \
-    save_notebook
+from .utils import colab_incompatible, grade_zip_file, grading_mode_disabled, list_available_tests, \
+    logs_event, resolve_test_info, running_on_colab, save_notebook
 from ..execute import Checker
 from ..export import export_notebook
 from ..plugins import PluginCollection
@@ -30,78 +30,6 @@ from ..test_files.metadata_test import NOTEBOOK_METADATA_KEY
 _OTTER_LOG_FILENAME = ".OTTER_LOG"
 _SHELVE = False
 _ZIP_NAME_FILENAME = "__zip_filename__"
-
-
-class _NotebookDecorators:
-    """
-    A class housing static methods that serve as decorators for the ``Notebook`` class.
-    """
-
-    # TODO: maintain wrapped function docstrings
-
-    @staticmethod
-    def colab_incompatible(f):
-        """
-        A decator that raises an error if the wrapped function is called in an environment running on
-        Google Colab.
-        """
-        def colab_only_method(self, *args, **kwargs):
-            if self._colab:
-                raise RuntimeError("This method is not compatible with Google Colab")
-            return f(self, *args, **kwargs)
-        return colab_only_method
-
-    @staticmethod
-    def grading_mode_disabled(f):
-        """
-        A decorator that returns without calling the wrapped function if the Notebook grading mode
-        is enabled.
-        """
-        def non_grading_mode_method(self, *args, **kwargs):
-            if type(self)._grading_mode:
-                return
-            return f(self, *args, **kwargs)
-        return non_grading_mode_method
-
-    @staticmethod
-    def logs_event(event_type):
-        """
-        A decorator that ensures each call is logged in the Otter log with type ``event_type``.
-
-        Events logging a ``EventType.CHECK`` should return a 3-tuple of the question name, the
-        ``TestFile`` and an environment to shelve. All other methods should just return their 
-        default return value, which will be logged.
-        """
-
-        def event_logger(f):
-            """
-            Wraps a function and ensures that the call's results are logged using 
-            ``Notebook._log_event``.
-            """
-
-            def run_function(self, *args, **kwargs):
-                """
-                Runs a method, catching any errors and logging the call. Returns the return value
-                of the function, unless ``EventType.CHECK`` is used, in which case the return value
-                is assumed to be a 3-tuple and the second value in the tuple is returned.
-                """
-                try:
-                    if event_type == EventType.CHECK:
-                        question, results, shelve_env = f(self, *args, **kwargs)
-                    else:
-                        results = f(self, *args, **kwargs)
-                        shelve_env = {}
-                        question = None
-                except Exception as e:
-                    self._log_event(event_type, success=False, error=e)
-                    raise e
-                else:
-                    self._log_event(event_type, results=results, question=question, shelve_env=shelve_env)
-                return results
-
-            return run_function
-
-        return event_logger
 
 
 class Notebook:
@@ -119,7 +47,7 @@ class Notebook:
     _tests_dir_override = None
     _grading_mode = False
 
-    @_NotebookDecorators.logs_event(EventType.INIT)
+    @logs_event(EventType.INIT)
     def __init__(self, nb_path=None, tests_dir="./tests", colab=None):
         global _SHELVE
 
@@ -235,7 +163,7 @@ class Notebook:
 
         return nb_path
 
-    @_NotebookDecorators.logs_event(EventType.CHECK)
+    @logs_event(EventType.CHECK)
     def check(self, question, global_env=None):
         """
         Runs tests for a specific question against a global environment. If no global environment
@@ -269,7 +197,7 @@ class Notebook:
 
         return question, result, global_env
 
-    @_NotebookDecorators.colab_incompatible
+    @colab_incompatible
     def run_plugin(self, plugin_name, *args, nb_path=None, **kwargs):
         """
         Runs the plugin ``plugin_name`` with the specified arguments. Use ``nb_path`` if the path
@@ -291,9 +219,9 @@ class Notebook:
             self._plugin_collections[plugin_name] = pc
         pc.run("from_notebook", *args, **kwargs)
 
-    @_NotebookDecorators.grading_mode_disabled
-    @_NotebookDecorators.colab_incompatible
-    @_NotebookDecorators.logs_event(EventType.TO_PDF)
+    @grading_mode_disabled
+    @colab_incompatible
+    @logs_event(EventType.TO_PDF)
     def to_pdf(self, nb_path=None, filtering=True, pagebreaks=True, display_link=True, force_save=False):
         """
         Exports a notebook to a PDF using Otter Export
@@ -330,8 +258,8 @@ class Notebook:
             display(HTML(out_html))
         self._log_event(EventType.TO_PDF)
 
-    @_NotebookDecorators.grading_mode_disabled
-    @_NotebookDecorators.colab_incompatible
+    @grading_mode_disabled
+    @colab_incompatible
     def add_plugin_files(self, plugin_name, *args, nb_path=None, **kwargs):
         """
         Runs the ``notebook_export`` event of the plugin ``plugin_name`` and tracks the file paths
@@ -355,9 +283,9 @@ class Notebook:
             return
         self._addl_files.extend(addl_files)
 
-    @_NotebookDecorators.grading_mode_disabled
-    @_NotebookDecorators.colab_incompatible
-    @_NotebookDecorators.logs_event(EventType.END_EXPORT)
+    @grading_mode_disabled
+    @colab_incompatible
+    @logs_event(EventType.END_EXPORT)
     def export(self, nb_path=None, export_path=None, pdf=True, filtering=True, pagebreaks=True, files=[], 
             display_link=True, force_save=False, run_tests=False):
         """
@@ -449,8 +377,8 @@ class Notebook:
 
             display(HTML(out_html))
 
-    @_NotebookDecorators.grading_mode_disabled
-    @_NotebookDecorators.logs_event(EventType.END_CHECK_ALL)
+    @grading_mode_disabled
+    @logs_event(EventType.END_CHECK_ALL)
     def check_all(self):
         """
         Runs all tests on this notebook. Tests are run against the current global environment, so any
