+++ otter/check/utils.py
@@ -6,6 +6,7 @@ import json
 import os
 import tempfile
 import time
+import wrapt
 
 from glob import glob
 from IPython import get_ipython
@@ -87,6 +88,64 @@ def running_on_colab():
         ``bool``: whether the current environment is on Google Colab
     """
     return "google.colab" in str(get_ipython())
+
+
+@wrapt.decorator
+def colab_incompatible(wrapped, self, args, kwargs):
+    """
+    A decator that raises an error if the wrapped function is called in an environment running on
+    Google Colab.
+    """
+    if self._colab:
+        raise RuntimeError("This method is not compatible with Google Colab")
+    return wrapped(*args, **kwargs)
+
+
+@wrapt.decorator
+def grading_mode_disabled(wrapped, self, args, kwargs):
+    """
+    A decorator that returns without calling the wrapped function if the Notebook grading mode
+    is enabled.
+    """
+    if type(self)._grading_mode:
+        return
+    return wrapped(*args, **kwargs)
+
+
+def logs_event(event_type):
+    """
+    A decorator that ensures each call is logged in the Otter log with type ``event_type``.
+
+    Events logging a ``EventType.CHECK`` should return a 3-tuple of the question name, the
+    ``TestFile`` and an environment to shelve. All other methods should just return their 
+    default return value, which will be logged.
+    """
+    @wrapt.decorator        
+    def event_logger(wrapped, self, args, kwargs):
+        """
+        Runs a method, catching any errors and logging the call. Returns the return value
+        of the function, unless ``EventType.CHECK`` is used, in which case the return value
+        is assumed to be a 3-tuple and the second value in the tuple is returned.
+        """
+        try:
+            if event_type == EventType.CHECK:
+                question, results, shelve_env = wrapped(*args, **kwargs)
+
+            else:
+                results = wrapped(*args, **kwargs)
+                shelve_env = {}
+                question = None
+
+        except Exception as e:
+            self._log_event(event_type, success=False, error=e)
+            raise e
+
+        else:
+            self._log_event(event_type, results=results, question=question, shelve_env=shelve_env)
+
+        return results
+
+    return event_logger
 
 
 def list_available_tests(tests_dir, nb_path):
